pyproject.toml
[tool.poetry]
name = "louis-project"
version = "0.1.0"
description = "No description"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.main"
